rendering types names on submit
